ENG-18870-ap-bigj-update2:
Modified Benchmark.java & UpdateClassesThread.java, to randomly select
which txnid*.jar file to use when updating, including the bigjar files.
For now, for test purposes, bigjarratio is set to 0.50 (i.e. 50%), so
they will get used a lot; but bigjarjavaratio is 0.0, so the bigjar
files containing lots of Java classes will not be used. (I'll change
bigjarjavaratio in a later test.) [This time, based on yesterday's
(2/19) master.]